+++ core/src/test/java/feign/client/DefaultClientTest.java
@@ -50,7 +50,7 @@ public class DefaultClientTest {
     @RequestLine("POST /?foo=bar&foo=baz&qux=")
     @Headers({"Foo: Bar", "Foo: Baz", "Qux: ", "Content-Type: text/plain"}) Response post(String body);
 
-    @RequestLine("PATCH /") String patch();
+    @RequestLine("PATCH /") @Headers("Accept: text/plain") String patch();
   }
 
   @Test public void parsesRequestAndResponse() throws IOException, InterruptedException {
@@ -69,7 +69,7 @@ public class DefaultClientTest {
 
     assertThat(server.takeRequest()).hasMethod("POST")
         .hasPath("/?foo=bar&foo=baz&qux=")
-        .hasHeaders("Foo: Bar", "Foo: Baz", "Qux: ", "Content-Length: 3")
+        .hasHeaders("Foo: Bar", "Foo: Baz", "Qux: ", "Accept: */*", "Content-Length: 3")
         .hasBody("foo");
   }